Move classes to gov.nci namespace.
Addresses #56
A number of classes were mistakenly put in the com.nci namespace instead
of gov.nci.
