+++ resources/testng-CTS.xml
@@ -9,10 +9,10 @@
 			</run>
 		</groups>
 		<classes> 	
-			  <class name="com.nci.testcases.BasicSearch_Test" />  
-			   <!-- <class name="com.nci.testcases.AdvanceSearch_Test" /> 
-			    <class name="com.nci.testcases.AdvanceSearchResults_Test" />    		
-			<class name="com.nci.testcases.BasicSearchResults_Test" />-->
+			  <class name="gov.nci.testcases.BasicSearch_Test" />  
+			   <!-- <class name="gov.nci.testcases.AdvanceSearch_Test" /> 
+			    <class name="gov.nci.testcases.AdvanceSearchResults_Test" />    		
+			<class name="gov.nci.testcases.BasicSearchResults_Test" />-->
 			    
 		</classes>
 	</test> 
@@ -25,11 +25,11 @@
 			</run>
 		</groups>
 		<classes> 	
-			 <class name="com.nci.testcases.BasicSearch_Test" /> 
-			<class name="com.nci.testcases.AdvanceSearch_Test" />
+			 <class name="gov.nci.testcases.BasicSearch_Test" /> 
+			<class name="gov.nci.testcases.AdvanceSearch_Test" />
 			
 			 
-			  <class name="com.nci.testcases.AdvanceSearchResults_Test" />  
+			  <class name="gov.nci.testcases.AdvanceSearchResults_Test" />  
 			
 		</classes>
 	</test>  -->
@@ -42,10 +42,10 @@
 			</run>
 		</groups>
 		<classes> 	
-		 <class name="com.nci.testcases.BasicSearch_Test" /> 
-		<class name="com.nci.testcases.AdvanceSearch_Test" />
+		 <class name="gov.nci.testcases.BasicSearch_Test" /> 
+		<class name="gov.nci.testcases.AdvanceSearch_Test" />
 			
-		 <class name="com.nci.testcases.AdvanceSearchResults_Test" /> 
+		 <class name="gov.nci.testcases.AdvanceSearchResults_Test" /> 
 			
 		</classes>
 	</test> -->
@@ -57,8 +57,8 @@
 			</run>
 		</groups>
 		<classes> 	
-			<class name="com.nci.testcases.BasicSearch_Test" />
-		<class name="com.nci.testcases.AdvanceSearch_Test" />
+			<class name="gov.nci.testcases.BasicSearch_Test" />
+		<class name="gov.nci.testcases.AdvanceSearch_Test" />
 		</classes>
 	</test>  -->
 	<!--   <test name="NCI - Firefox Testing" >
@@ -69,8 +69,8 @@
 			</run>
 		</groups>
 		<classes> 	
-			 <class name="com.nci.testcases.BasicSearch_Test" /> 
-			   <class name="com.nci.testcases.AdvanceSearch_Test" />
+			 <class name="gov.nci.testcases.BasicSearch_Test" /> 
+			   <class name="gov.nci.testcases.AdvanceSearch_Test" />
 		</classes>
 	</test>  --> 
 	<!-- <test name="NCI - Firefox Testing Headless" >
@@ -81,8 +81,8 @@
 			</run>
 		</groups>
 		<classes> 	
-			<class name="com.nci.testcases.BasicSearch_Test" />
-			  <class name="com.nci.testcases.AdvanceSearch_Test" /> 
+			<class name="gov.nci.testcases.BasicSearch_Test" />
+			  <class name="gov.nci.testcases.AdvanceSearch_Test" /> 
 		</classes>
 	</test>   -->
 </suite>
