+++ src/main/java/gov/nci/Utilities/BrowserManager.java
@@ -1,4 +1,4 @@
-package com.nci.Utilities;
+package gov.nci.Utilities;
 
 import java.util.HashMap;
 import java.util.Map;
