+++ src/main/java/gov/nci/Utilities/ScreenShot.java
@@ -1,4 +1,4 @@
-package com.nci.Utilities;
+package gov.nci.Utilities;
 
 import java.io.File;
