+++ src/main/java/gov/nci/clinicalTrial/pages/BasicSearch.java
@@ -1,4 +1,4 @@
-package com.nci.clinicalTrial.pages;
+package gov.nci.clinicalTrial.pages;
 
 import org.openqa.selenium.Keys;
 import org.openqa.selenium.WebDriver;
@@ -7,7 +7,7 @@ import org.openqa.selenium.support.FindBy;
 import org.openqa.selenium.support.How;
 import org.openqa.selenium.support.PageFactory;
 
-import com.nci.Utilities.FunctionLibrary;
+import gov.nci.Utilities.FunctionLibrary;
 
 public class BasicSearch {
