+++ src/test/java/gov/nci/testcases/BaseClass.java
@@ -1,4 +1,4 @@
-package com.nci.testcases;
+package gov.nci.testcases;
 
 import java.text.SimpleDateFormat;
 import java.util.Date;
@@ -16,8 +16,9 @@ import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.BeforeTest;
 import org.testng.annotations.Parameters;
 
-import com.nci.Utilities.ConfigReader;
-import com.nci.Utilities.ScreenShot;
+import gov.nci.Utilities.ConfigReader;
+import gov.nci.Utilities.ScreenShot;
+
 import com.relevantcodes.extentreports.ExtentReports;
 import com.relevantcodes.extentreports.ExtentTest;
 import com.relevantcodes.extentreports.LogStatus;
